requirements.txt
asttokens==2.0.3
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
certifi==2019.3.9
chardet==3.0.4
cytoolz==0.10.1
Django==2.2.13
django-chunked-upload==1.1.3
django-crispy-forms==1.7.2
django-filter==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.6.0
ipfshttpclient==0.4.13.2
jdcal==1.4
jsonschema==3.2.0
lru-dict==1.1.6
mariasql==5
Markdown==3.2.2
multiaddr==0.0.9
mysqlclient==1.4.2.post1
netaddr==0.7.19
openpyxl==2.6.2
parsimonious==0.8.1
Pillow==7.1.0
protobuf==3.11.3
pycryptodome==3.9.7
PyMySQL==0.9.3
pyrsistent==0.16.0
pytz==2018.4
requests==2.21.0
requests-toolbelt==0.9.1
rlp==1.2.0
six==1.14.0
sqlparse==0.3.1
style==1.1.0
toolz==0.10.0
typing-extensions==3.7.4.2
update==0.0.1
urllib3==1.24.2
varint==1.0.2
vyper==0.2.3
web3==5.8.0
websockets==8.1
zipp==3.1.0
